Moving config to directory
