+++ Gruntfile.js
@@ -132,6 +132,12 @@ module.exports = function (grunt) {
           },
           {
             expand: true,
+            cwd: 'test/e2e/appium',
+            src: 'ios-conf.js',
+            dest: 'target/e2e/appium'
+          },
+          {
+            expand: true,
             cwd: 'test/e2e/specs/',
             src: '*',
             dest: 'target/e2e/specs/'
