+++ scripts/travis.sh
@@ -7,9 +7,7 @@ if [ "${TRAVIS_EVENT_TYPE}" = "cron" ] ; then
     yarn test:e2e
 else
     yarn build:release
-    pwd
-    ls ./test/e2e
-    # yarn lint
-    # yarn test
-    # yarn test:e2e
+    yarn lint
+    yarn test
+    yarn test:e2e
 fi
